requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
umap-learn>=0.5.6
hdbscan>=0.8.33
scikit-learn>=1.4.0
numpy>=1.26.0
anthropic>=0.30.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
respx>=0.21.0
